+++ examples/scripts/battery_parameterisation/tau_reparameterised_ecm.py
@@ -115,7 +115,7 @@ print(
 print("Estimated parameters:", results.x.tolist() + [results.x[2] / results.x[1]])
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)