+++ examples/scripts/comparison_examples/exponential_decay.py
@@ -4,11 +4,8 @@ import pybamm
 import pybop
 
 # Define model and use high-performant solver for sensitivities
-solver = pybamm.IDAKLUSolver
 parameter_set = pybamm.ParameterValues({"k": 1, "y0": 0.5})
-model = pybop.ExponentialDecayModel(
-    parameter_set=parameter_set, solver=solver, n_states=2
-)
+model = pybop.ExponentialDecayModel(parameter_set=parameter_set, n_states=2)
 
 # Fitting parameters
 parameters = pybop.Parameters(
@@ -59,7 +56,7 @@ optim = pybop.AdamW(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)