+++ examples/scripts/getting_started/mcmc_example.py
@@ -1,10 +1,13 @@
+import sys
+
 import numpy as np
-import pybamm
 
 import pybop
 
+# Set parallelization if on macOS / Unix
+parallel = True if sys.platform != "win32" else False
+
 # Parameter set and model definition
-solver = pybamm.IDAKLUSolver()
 parameter_set = pybop.ParameterSet("Chen2020")
 parameter_set.update(
     {
@@ -12,7 +15,7 @@ parameter_set.update(
         "Positive electrode active material volume fraction": 0.71,
     }
 )
-synth_model = pybop.lithium_ion.DFN(parameter_set=parameter_set, solver=solver)
+synth_model = pybop.lithium_ion.SPMe(parameter_set=parameter_set)
 
 # Fitting parameters
 parameters = pybop.Parameters(
@@ -33,7 +36,7 @@ init_soc = 0.5
 sigma = 0.005
 experiment = pybop.Experiment(
     [
-        ("Discharge at 0.5C for 6 minutes (5 second period)",),
+        ("Discharge at 0.5C for 3 minutes (5 second period)",),
     ]
 )
 values = synth_model.predict(
@@ -51,28 +54,26 @@ dataset = pybop.Dataset(
         "Time [s]": values["Time [s]"].data,
         "Current function [A]": values["Current [A]"].data,
         "Voltage [V]": noisy(values["Voltage [V]"].data, sigma),
-        "Bulk open-circuit voltage [V]": noisy(
-            values["Bulk open-circuit voltage [V]"].data, sigma
-        ),
     }
 )
 
-model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=pybamm.IDAKLUSolver())
-model.build(initial_state={"Initial SoC": init_soc})
-signal = ["Voltage [V]", "Bulk open-circuit voltage [V]"]
+model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
+signal = ["Voltage [V]"]
 
 # Generate problem, likelihood, and sampler
-problem = pybop.FittingProblem(model, parameters, dataset, signal=signal)
+problem = pybop.FittingProblem(
+    model, parameters, dataset, signal=signal, initial_state={"Initial SoC": init_soc}
+)
 likelihood = pybop.GaussianLogLikelihood(problem)
 posterior = pybop.LogPosterior(likelihood)
 
 sampler = pybop.DifferentialEvolutionMCMC(
     posterior,
     chains=3,
-    max_iterations=300,
+    max_iterations=250,  # Reduced for CI, increase for improved posteriors
     warm_up=100,
     verbose=True,
-    # parallel=True,  # uncomment to enable parallelisation (macOS/WSL/Linux only)
+    parallel=parallel,  # (macOS/WSL/Linux only)
 )
 chains = sampler.run()