+++ pybop/optimisers/_result.py
@@ -162,7 +162,9 @@ class OptimisationResult:
             ValueError: If the optimised parameters do not produce a finite cost value.
         """
         if not any(np.isfinite(self._final_cost)):
-            raise ValueError("Optimised parameters do not produce a finite cost value")
+            raise ValueError(
+                f"Optimised parameters {self.cost.parameters.as_dict()} do not produce a finite cost value"
+            )
 
     def check_physical_viability(self, x):
         """
@@ -207,7 +209,6 @@ class OptimisationResult:
         # Format the sensitivities
         self.sense_format = ""
         if self._sensitivities:
-            self.sense_format = ""
             for value, conf in zip(
                 self._sensitivities["ST"], self._sensitivities["ST_conf"]
             ):