+++ tests/integration/test_thevenin_parameterisation.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pytest
-from pybamm import IDAKLUSolver
 
 import pybop
 
@@ -78,11 +77,10 @@ class TestTheveninParameterisation:
             (pybop.PSO, ""),
         ],
     )
-    def test_optimisers_on_simple_model(
+    def test_optimisers_on_thevenin_model(
         self, model, parameters, dataset, cost_class, optimiser, method
     ):
         # Define the cost to optimise
-        model.solver = IDAKLUSolver()
         problem = pybop.FittingProblem(model, parameters, dataset)
         cost = cost_class(problem)
 
@@ -127,5 +125,5 @@ class TestTheveninParameterisation:
                 ),
             ]
         )
-        sim = model.predict(experiment=experiment)
-        return sim
+        sol = model.predict(experiment=experiment)
+        return sol