+++ tests/unit/test_plots.py
@@ -1,7 +1,6 @@
 import warnings
 
 import numpy as np
-import pybamm
 import pytest
 from packaging import version
 
@@ -83,7 +82,6 @@ class TestPlots:
 
     @pytest.fixture
     def jax_fitting_problem(self, model, parameters, dataset):
-        model.solver = pybamm.IDAKLUSolver()
         problem = pybop.FittingProblem(model, parameters, dataset)
         problem.model.jaxify_solver(t_eval=problem.domain_data)
         return problem
@@ -103,12 +101,12 @@ class TestPlots:
 
     def test_problem_plots(self, fitting_problem, design_problem, jax_fitting_problem):
         # Test plot of Problem objects
-        pybop.plot.quick(fitting_problem, title="Optimised Comparison")
-        pybop.plot.quick(design_problem)
-        pybop.plot.quick(jax_fitting_problem)
+        pybop.plot.problem(fitting_problem, title="Optimised Comparison")
+        pybop.plot.problem(design_problem)
+        pybop.plot.problem(jax_fitting_problem)
 
         # Test conversion of values into inputs
-        pybop.plot.quick(fitting_problem, problem_inputs=[0.6, 0.6])
+        pybop.plot.problem(fitting_problem, problem_inputs=[0.6, 0.6])
 
     @pytest.fixture
     def cost(self, fitting_problem):